refactor(codec): split interface for Go-style flexibility and extensibility
