+++ stdwrap/gson/gson_codec.go
@@ -1,14 +1,14 @@
 // Copyright 2025 Bytedance Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance By the License.
+// you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
 //     http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
-// ByOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
@@ -31,26 +31,51 @@
 //
 //   - Validation: [ValidBy]
 //   - Marshal to []byte: [MarshalBy], [MarshalIndentBy]
-//   - Marshal to string: [MarshalToStringBy], [ToStringBy], [ToStringIndentBy]
+//   - Marshal to string: [MarshalStringBy], [ToStringBy], [ToStringIndentBy]
 //   - Unmarshal to object: [UnmarshalBy]
 package gson
 
 import (
-	"encoding/json"
-
 	"github.com/bytedance/gg/internal/conv"
 )
 
-// Codec abstracts encoding and decoding of values.
-type Codec interface {
+// Marshaler defines the interface for serializing a value into a byte slice.
+type Marshaler interface {
 	Marshal(v any) ([]byte, error)
-	MarshalIndent(v any, prefix, indent string) ([]byte, error)
+}
+
+// Unmarshaler defines the interface for deserializing data into a Go value.
+type Unmarshaler interface {
 	Unmarshal(data []byte, out any) error
+}
+
+// PrettyMarshaler defines the interface for pretty-printing serialized output with indentation.
+type PrettyMarshaler interface {
+	MarshalIndent(v any, prefix, indent string) ([]byte, error)
+}
+
+// Validator defines the interface for validating whether a byte slice is a valid-encoded format.
+type Validator interface {
 	Valid(data []byte) bool
 }
 
+// Codec is the minimal interface for encoding and decoding values.
+// It combines Marshaler and Unmarshaler.
+type Codec interface {
+	Marshaler
+	Unmarshaler
+}
+
+// FullCodec is an extended interface for codecs that support indentation and validation.
+// It combines Codec, PrettyMarshaler, and Validator.
+type FullCodec interface {
+	Codec
+	PrettyMarshaler
+	Validator
+}
+
 // ValidBy reports whether the input data is valid according to the given codec.
-func ValidBy[V ~[]byte | ~string](codec Codec, data V) bool {
+func ValidBy[V ~[]byte | ~string](codec FullCodec, data V) bool {
 	switch v := any(data).(type) {
 	case string: // support types like ~string
 		return codec.Valid(conv.StringToBytes(v))
@@ -75,18 +100,18 @@ func MarshalBy[T any](codec Codec, v T) ([]byte, error) {
 }
 
 // MarshalIndentBy marshals the value v into indented bytes using the provided codec.
-func MarshalIndentBy[T any](codec Codec, v T, prefix, indent string) ([]byte, error) {
+func MarshalIndentBy[T any](codec FullCodec, v T, prefix, indent string) ([]byte, error) {
 	return codec.MarshalIndent(v, prefix, indent)
 }
 
-// MarshalToStringBy marshals the value v into a JSON string using the provided codec.
+// MarshalStringBy marshals the value v into a JSON string using the provided codec.
 //
 // 🚀 Example:
 //
-//	MarshalToStringBy(codec, map[string]any{"name": "test", "age": 10}) ⏩  "{\"name\":\"test\",\"age\":10}"
+//	MarshalStringBy(codec, map[string]any{"name": "test", "age": 10}) ⏩  "{\"name\":\"test\",\"age\":10}"
 //
 // 💡 HINT: For high-performance JSON serialization, see [github.com/json-iterator/go] or [github.com/bytedance/sonic].
-func MarshalToStringBy[T any](codec Codec, v T) (string, error) {
+func MarshalStringBy[T any](codec Codec, v T) (string, error) {
 	data, err := codec.Marshal(v)
 	return conv.BytesToString(data), err
 }
@@ -98,7 +123,7 @@ func ToStringBy[T any](codec Codec, v T) string {
 }
 
 // ToStringIndentBy returns the indented string representation of v using the codec, ignoring errors.
-func ToStringIndentBy[T any](codec Codec, v T, prefix, indent string) string {
+func ToStringIndentBy[T any](codec FullCodec, v T, prefix, indent string) string {
 	data, _ := codec.MarshalIndent(v, prefix, indent)
 	return conv.BytesToString(data)
 }
@@ -115,24 +140,3 @@ func UnmarshalBy[T any, V ~[]byte | ~string](codec Codec, v V) (T, error) {
 	err := codec.Unmarshal([]byte(v), &t)
 	return t, err
 }
-
-// default json std lib
-type stdCodec struct{}
-
-func (stdCodec) Marshal(v any) ([]byte, error) {
-	return json.Marshal(v)
-}
-
-func (stdCodec) MarshalIndent(v any, prefix, indent string) ([]byte, error) {
-	return json.MarshalIndent(v, prefix, indent)
-}
-
-func (stdCodec) Unmarshal(data []byte, out any) error {
-	return json.Unmarshal(data, out)
-}
-
-func (stdCodec) Valid(data []byte) bool {
-	return json.Valid(data)
-}
-
-var JsonStdCodec Codec = stdCodec{}
